requirements.txt
cffi==1.14.6
cryptography==35.0.0
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
aiohttp==3.8.3
asyncio==3.4.3